requirements.txt
flask
flask-cors
pandas
pywebview
